Move classes and remove unnecessary class. Add PostgresChangeBuilder#all
