+++ .gitignore
@@ -14,6 +14,6 @@
 /Postgrest/build/
 /Storage/build/
 /Realtime/build/
-/Realtime/src/desktopMain/Test.kt
+/Realtime/src/desktopMain/kotlin/Test.kt
 .idea
 local.properties
