+++ Realtime/build.gradle.kts
@@ -52,6 +52,7 @@ kotlin {
             dependencies {
                 //logback-classic
                 // https://mvnrepository.com/artifact/ch.qos.logback/logback-classic
+
             }
         }
         val androidMain by getting
