--- Realtime/src/commonMain/kotlin/io/github/jan/supacompose/realtime/events/EventListener.kt
@@ -1,15 +1,0 @@
-package io.github.jan.supacompose.realtime.events
-
-import io.github.jan.supacompose.realtime.events.actions.ChannelAction
-
-/**
- * An event listener for [ChannelAction]s
- */
-fun interface EventListener {
-
-    /**
-     * Called whenever a [ChannelAction] is received
-     */
-    fun onEvent(event: ChannelAction)
-
-}
